+++ Domain/Ports/Infrastructure/IAltinnRecoveryService.cs
@@ -1,4 +1,6 @@
 using Arbeidstilsynet.Common.Altinn.Model.Adapter;
+using Arbeidstilsynet.Common.Altinn.Model.Api.Response;
+using Arbeidstilsynet.MeldingerReceiver.Domain.Ports.Infrastructure.Dto;
 
 namespace Arbeidstilsynet.MeldingerReceiver.Domain.Ports.Infrastructure;
 
@@ -13,4 +15,17 @@ public interface IAltinnRecoveryService
     Task<
         Dictionary<string, IEnumerable<AltinnMetadata>>
     > GetMetadataForAllNonCompletedInstancesForRegisteredApps();
+
+    Task<AltinnInstance?> GetInstanceMetadata(Guid instanceGuid, CancellationToken ct = default);
+
+    Task<IReadOnlyList<DataElement>?> GetDataElementsForInstance(
+        Guid instanceGuid,
+        CancellationToken ct = default
+    );
+
+    Task<DataElementDownload?> DownloadDataElement(
+        Guid instanceGuid,
+        Guid dataElementId,
+        CancellationToken ct = default
+    );
 }